Assume git server is already configured and be more automated
Prior to this commit, I was trying to allow someone to spin up
everything from scratch and have it magically work.  That's
significantly more work that I haven't completed.

So in the short term...

After this commit, we'll assume a one-time setup cost of your
git server and then from there on out will just assume it's setup
correctly and should be referred to from code manager.